AbyssalCraft 1.9.4.10
*A bunch of textures have been revamped (courtesy of Dylan4ever)
*Lesser Shoggoths can no longer break Bedrock (or other unbreakable
blocks) with their acid
*Acid Projectiles no longer break blocks with Tile Entities when hitting
Players or other mobs
*Scaled down the model of the baby Lesser Shoggoth (to be more
proportional with the hitbox, also made the eyes a bit bigger)
*Slowed down the initial velocity of the Acid Projectile, increased the
time between each projectile being fired, and made it so baby Lesser
Shoggoths won't fire them
*Added a config option for changing the frequency at which Lesser
Shoggoths spit acid
*J'zahar now respects the mobGriefing gamerule with his explosion
(closes #292 )
*Fixed a dupe bug in the Materializer GUI
*Mobs spawned through a summoning ritual now have their portal cooldown
set to max
*Increased the time it takes for a Greater Dread Spawn and Lesser
Dreadbeast to spawn a regular Dread Spawn
*Increased the time it takes for Cha'garoth to spawn the various things
he spawns, and removed the direct physical damage dealt by his fireballs
*Reduced the range of Cha'garoth's melee attacks
*Fixed the pick block for Calcified Stone

+++ src/main/java/com/shinoow/abyssalcraft/common/entity/EntityAcidProjectile.java
@@ -107,7 +107,8 @@ public class EntityAcidProjectile extends EntityThrowable {
 									BlockPos pos2 = new BlockPos(xp, yp, zp);
 									if(!world.isAirBlock(pos2) && world.getBlockState(pos2).getBlockHardness(world, pos2) < ACConfig.acidResistanceHardness && world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_ooze
 											&& world.getBlockState(pos2) != ACBlocks.stone.getDefaultState().withProperty(BlockACStone.TYPE, EnumStoneType.MONOLITH_STONE)
-											&& world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos).getBlock().hasTileEntity(world.getBlockState(pos)))
+											&& world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos2).getBlock().hasTileEntity(world.getBlockState(pos2))
+											&& world.getBlockState(pos2).getBlockHardness(world, pos2) != -1 && world.getBlockState(pos2).getBlock().canEntityDestroy(world.getBlockState(pos2), world, pos2, getThrower()))
 										world.destroyBlock(pos2, false);
 								}
 						}
@@ -121,7 +122,8 @@ public class EntityAcidProjectile extends EntityThrowable {
 				BlockPos pos = mop.getBlockPos();
 				if(!world.isAirBlock(pos) && world.getBlockState(pos).getBlockHardness(world, pos) < ACConfig.acidResistanceHardness && world.getBlockState(pos).getBlock() != ACBlocks.shoggoth_ooze
 						&& world.getBlockState(pos) != ACBlocks.stone.getDefaultState().withProperty(BlockACStone.TYPE, EnumStoneType.MONOLITH_STONE)
-						&& world.getBlockState(pos).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos).getBlock().hasTileEntity(world.getBlockState(pos)))
+						&& world.getBlockState(pos).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos).getBlock().hasTileEntity(world.getBlockState(pos))
+						&& world.getBlockState(pos).getBlockHardness(world, pos) != -1 && world.getBlockState(pos).getBlock().canEntityDestroy(world.getBlockState(pos), world, pos, getThrower()))
 					world.destroyBlock(pos, false);
 				setDead();
 			}
